refactor: JWT 생성 및 인증/인가 로직 구조 개선 - 토큰에 고유 식별자(id), 역할(role) 정보만 포함 - LoginMember, LoginUserArgumentResolver roomescape 내부로 이동
