+++ src/main/java/roomescape/config/WebConfig.java
@@ -1,6 +1,6 @@
 package roomescape.config;
 
-import auth.ui.LoginUserArgumentResolver;
+import roomescape.auth.LoginUserArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
